split out mono repo and working docker compose
