+++ .nvmrc
@@ -1,0 +1,1 @@
+v19.8.1
