+++ client/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "crypto",
+	"name": "crypto-client",
 	"private": true,
 	"version": "0.0.0",
 	"type": "module",
