+++ client/src/__mocks__/svg.tsx
