+++ client/src/components/footer/footer.test.tsx
