+++ client/src/components/footer/footer.tsx
