+++ client/tsconfig.json
@@ -16,7 +16,7 @@
     "jsx": "react-jsx",
     "esModuleInterop": true,
   },
-  "include": ["src"],
+  "include": ["src", "client/src/client", "server/src/server", "client/src/__mocks__", "server/src/routes"],
   "ts-node": {
     "esm": true
   },
