+++ docker-compose.yml
@@ -1,27 +1,23 @@
 version: '3.8'
 
 services:
+  api:
+    build: ./server
+    container_name: api_backend
+    ports:
+      - 3000:3000
+    volumes:
+      - ./server:/api
+      - /api/node_modules
+    stdin_open: true
+    tty: true
   client:
-    build: 
-      context: .
-      dockerfile: ./src/client/Dockerfile
+    build: ./client 
     container_name: client_frontend
     ports:
       - 8000:8000
     volumes:
-      - ./src/client:/app
-      - ./app/node_modules
-    stdin_open: true
-    tty: true
-  api:
-    build: 
-      context: .
-      dockerfile: ./src/server/Dockerfile
-    container_name: api_backend
-    ports:
-      - 5000:5000
-    volumes:
-      - ./src/server:/app
-      - ./app/node_modules
+      - ./client/:/app
+      - /app/node_modules
     stdin_open: true
     tty: true
