+++ server/package.json
@@ -1,0 +1,33 @@
+{
+	"name": "crypto-api",
+	"private": true,
+	"version": "0.0.0",
+	"type": "module",
+	"scripts": {
+		"start:server": "nodemon --esm --experimental-specifier-resolution=node src/server.ts",
+		"preview": "vite preview",
+		"test": "jest",
+		"lint": "tsc --noEmit && eslint src/**/*.ts{,x} --cache --max-warnings=0",
+		"lint:fix": "eslint src/**/*.ts{,x} --fix"
+	},
+	"dependencies": {
+		"@testing-library/jest-dom": "^5.16.5",
+		"@types/express": "^4.17.17",
+		"@types/jest": "^29.4.2",
+		"@types/supertest": "^2.0.12",
+		"axios": "^1.3.4",
+		"dotenv": "^16.0.3",
+		"express": "^4.18.2",
+		"jest": "^29.5.0",
+		"nodemon": "^2.0.21",
+		"supertest": "^6.3.3",
+		"ts-jest": "^29.0.5",
+		"ts-node": "^10.9.1"
+	},
+	"devDependencies": {
+		"autoprefixer": "^10.4.14",
+		"eslint": "^8.24.0",
+		"postcss": "^8.4.21",
+		"typescript": "^4.9.5"
+	}
+}
