+++ server/src/app.ts
@@ -1,5 +1,5 @@
 import express, { type Express, json } from 'express';
-import { routes } from './routes/';
+import { routes } from './routes';
 
 const app: Express = express();
