+++ server/src/routes/health-check.ts
