+++ server/src/routes/status.test.ts
