+++ server/src/routes/status.ts
